transfer some utils
